BUG#28061945 - NO_AUTO_CREATE_USER SQL-MODE SETTING PREVENTS 8.0 SERVER START ON WINDOWS.
No error is logged to the log file when server is started with sql mode of
NO_AUTO_CREATE_USER. This option is disallowed in 8.0 and is treated as
incorrect value in sql_mode and server is exited when this option is disallowed.
The fix is to ensure that when initialization of the variables fail, log the error
if log-error value is set to log-file.
This patch is provided by Daniel Blanchard and committed on his behalf.