Bug #22958077 ORDER BY LAST_SEEN_TRANSACTION RESULTS IN EMPTY SET (OR DEBUG               ASSERTION)
Problem:
--------
Replication P_S tables returns empty set when ORDER BY is
used.

Analysis:
---------
In replication P_S tables, the current position pointer to
the record was being used as an iterator for
fetching the records.

As a result, the current position
(table_replication_*::m_pos) of the record was getting
incremented in table_replication_*::rnd_next() and was
leading to PFS_engine_table::get_position() return the wrong
value to the Optimizer and hence was returning empty set.

Fix:
----
The current position pointer is made to point to the latest
record and is updated only when a new record is available.