Bug#24687073 - SET PASSWORD REPLICATES INCORRECTLY BETWEEN                5.6 AND 5.7
DESCRIPTION
===========
http://dev.mysql.com/doc/refman/5.7/en/set-password.html
documents that between 5.6 and 5.7, one format of
"SET PASSWORD" changes its syntax from:

SET PASSWORD [for user] = [hash of password]
to
SET PASSWORD [for user] = [password]

In cross version replication scenario when the master is in
5.6 and slave is in 5.7, the above SET PASSWORD command
results in a different password hash on latter than the
former.

ANALYSIS
========
Note that this bug can also be reproduced when both master
and slave belong to 5.7 with
'log-builtin-as-identified-by-password=ON'. In such a case
SET PASSWORD statements are logged as SET PASSWORD
statements, rather than being rewritten as ALTER USER
statements. So when the command:

SET PASSWORD [for user] = [password]

is fed to slave, it is hashing the already hashed input
string which results in master having hash of password and
slave having double hash of password.

Currently the flag 'uses_identified_by_clause' is set to
false in change_password() if
'log_builtin_as_identified_by_password' is true and it's a
slave thread. This flag is used in
set_and_validate_user_attributes() to determine if hashing
is required or not.

FIX
===
Relaxed this criteria by removing the 1st condition. So now
whenever its a slave thread (or a binlog applier thread),
it can be safely assumed that the input string is hash that
was already computed on master.

We immediately set 'uses_authentication_string_clause' to
true here so that it later avoids double hash in
set_and_validate_user_attributes()

The fix is restricted to 5.7 since we support cross-
version replication only for two consecutive versions
which in this case would be 5.6 and 5.7.