Bug #28112444	PB2 FAILURE OF PARTITION_NAMES.TEST IN WEEKLY-5.7 ON WINDOWS
PROBLEM
-------

This test is failing sporadically in windows because windows only
supports path length of 260 characters. When we run this test
with really long path and with long partition names it fails
with OS error saying "too long path" .

FIX
---
Presently disabling this test for windows. We should enable
NTFS long name feature in all pb2 machines and then enable
this test for windows.