BUG#28096808 - CONTRIBUTION BY FACEBOOK: DO NOT USE SIGMASK IN PPOLL FOR CLIENT LIBRARIES
Signals are unblocked when we call vio_io_wait function which uses
ppoll on client side. This is not required for the client side usage
of VIO API. Thus the patch disables the signal call in ppoll by passing
nullptr value for the case where thread_id in vio is not set.
The patch is contributed by Facebook Inc. and also fixes the similar bug#28075623.