Bug#27857833 PB2 RPL.RPL_MTS_SLAVE_PRESERVE_COMMIT_ORDER_DEADLOCK FAILING SPORADICALLY FEB 18
Problem:
Global debug on master was set to "d,set_commit_parent_100".This will make
transactions have same commit parent. So they can be applied parallel on slave.
Due to this slave was trying to execute CREATE TABLE t2 and INSERT INTO t2 parallely and hit ERROR 1146.

Solution:
SET debug=""; on master before CREATE and INSERT. This patch also includes 'start_slave.inc'
at the end to maintain the state of slave same as in the begining of test.