Fix an issue where the AutoTest of 'testFK -n CreateDropDuring' sometime crashed with an assert indicating that the indexes Vector was read beyond its end. However, inspecting the crash in a debugger shows that all accessed to the Vector were below its upper limit.
Reason turned out to be a few 'abort()'s which had been added
to the code in places where NDBT_FAILED was about to be returned.
Probably with the intention of creating a stack dump usable for
finding where the test failed. However, as 'indexes' was
declared as 'static':

static Vector<const NdbDictionary::Index*> indexes

.. The abort() caused the d'tor of the static Vector
to be called, effectively tearing away the indexes Vector
under the feets of the still running threads.

This patch remove these abort(), and replace it with more
extensive use of the CHK macros which will provide a
backtrack of where the failure happened.

Note that patch is not intended to fix AutoTest 'TEST' failures,
but should avoid that failures ends in mysterious core dumps
providing little usefull information.