Bug#27898384 INNODB: ASSERTION FAILURE: LOCK0LOCK.CC:4226:!(LOCK2->TRX == TRX && LOCK_GET_MOD
I separated row_prebuilt_t::new_rec_locks integer into two separate booleans,
to track not only how many locks we have created in RC mode, but also in
which particular index (primary or secondary) they are, so that we
unlock the proper locks when we no longer need them.
Also, since we only "describe" the lock we want to unlock (instead of
pointing to it directly) when calling lock_rec_unlock, it was crucial to
make the description more specific.
Thus, I narrowed the condition to search for LOCK_REC_NO_GAP
only (as opposed to INSERT INTENTION or other locks we might have
created in earlier queries).
There is a debug assert in the code which will inform us if there are
still cases in which this description matches more than one lock.
I've added the lock_rec_unlock.test to check lock_rec_unlock in RC mode.

Reviewed-by: Debarun Banerjee <debarun.banerjee@oracle.com>