Bug#27891472: FLUSH LOGS WITH NO LOG FILE PERMISSION LEADS HANG
Changing the log file acquires the log-lock. If the file cannot
be changed, we try to inform the user of this. To do this, we
call the error logger, which in turn waits to acquire the log-lock,
and thus hangs forever.

With this patch, we first release the log-lock in case of an
error, then try to write the message (to the old log, if any),
and finally re-acquire the lock as the caller expects to have
to release it.