Bug #27877485 GENERAL UNASSIGNED TABLESPACE CRASHES MYSQL IN AN UPGRADE
Problem:

When an empty general tablespace is created in 5.7 it will have
FIL_IBD_FILE_INITIAL_SIZE_5_7 (defined as 4) pages.  When this
tablespace is upgraded by running 8.0 server, then SDI index will be
created in this tablespace.  This will increase the tablespace size to
6.  This is still 1 less than what is expected in 8.0 which is
FIL_IBD_FILE_INITIAL_SIZE (defined as 7).

Solution:

Add a tolerance in the check so that the upgrade situation is taken
care of.

rb#19854 approved by Darshan.