BUG#28061910: ENABLE THE --TEST-PROGRESS OPTION BY DEFAULT
Description:
------------
When running multiple tests or tests from multiple suites, it is
necessary to specify the --test-progress option explicitly in order to
know the progress of the test run. It is good to know the current
progress by default, because at times user can make a choice to let
the current test run finish or abort it.

Fix:
----
Enabled test-progress option by default. To disable it, set the value
to 0 for 'test-progress' option.

Change-Id: Icaa2c1b4a5ffd80f3cab3913556174ee45d3d63d