Bug#27836543:STREAMING BINLOGS NOT WORKING ON ON WINDOWS
Problem:

Bug#27445278:CONTRIBUTION BY FACEBOOK: STREAMING BINLOGS was a contribution
bug from facebook, but the fix doesn't work on windows even after the
proposed changes.

The above bug has introduced a new method my_b_event_read which checks the
file descriptor of the stream from which it is reading and if the stream is
stdin it skips any magic byte which occurs in between the stream. But the
check to verify that a file descriptor belongs to standard input(stdin)or not
uses the linux specific method fileno(stdin) which will not work on
windows platform.

Fix:
Modify the file descriptor check to use mysql implementation of fileno() which
is my_fileno(file_descriptor *).