Bug#28165246 ADD SUPPORT FOR SOLARIS STUDIO 12.6 AKA 5.15
There are a few lines that fail to compile with:
Error: Use of overloaded binary operator '<<' is ambiguous.

There are a few files that fail to compile with:
Error: Use of overloaded binary operator '+=' is ambiguous.

X plugin tests fail with:
mysqltest: At line 13: Query 'INSTALL COMPONENT "file://component_mysqlx_global_reset"' failed.
This is because the function 'list_components' is hidden and not found by dlsym()

Use <var>.load() when printing atomic variables.

Add operator+=() for the failed/missing overload.

Move definition of list_components out of the anonymous namespace.

Change-Id: Ia73da591c80101820633a8231e5b54af11db8078