Bug #27419474 INNODB: ASSERTION FAILURE: ZLOB0READ.CC.*:FLST_LEN != 0
Problem:

When the LOB is purged or rollback, there was a crash after freeing
all the LOB index entries.  So while doing a crash recovery, when the
LOB was read, the index list was of length 0.  This is unexpected
behaviour.  If the LOB reference has length > 0, then the LOB index
entries are expected to be available.

Solution:

The index list becoming empty and the LOB reference being updated is
to be done in a single mtr.

rb#19892 approved by Krzysztof.