Bug#28170699: CRASH ON MACOS X WHEN LOWER_CASE_TABLE_NAMES=2 AND INNODB_FILE_PER_TABLE SET
The function Fil_path::is_valid_location() did not compare case insensitive when
lower_case_table_names=2.  This is the setting for MacOS where table names
are compared in lower case but schema directories are generated with the case
given in the DDL statement.  This comparison is done in order to enforce correct
locations of file-per-table tablespaces. This function and bug was introduced
with WL#8619.

Added a test case that uses mixed case identifiers in all thre setting of
--lower-case-table-names. (0,1 & 2)