Bug #25987758: HANDLER READ ON TABLE WITH COLUMN PRIVILEGES ASSERTS
The handler interface uses any_db for a database name to avoid
defaulting to the current database so it can handle aliases given in
handler open.

But the insert_fields iterator expects that the db name in the table_ref
and into the table are the same (except for I_S tables).
And in this case they are not.
Hence added any_db as a valid value for table_ref.db.
Added a test case.