Bug#27903881 - [MYSQL 8.0 GA RELEASE & DEBUG BUILD] FIL_SPACE_GET(TABLE->SPACE) != __NULL]
Issue
=====
For a partitioned table, FLUSH TABLES .... FOR EXPORT immediately
after a ALTER TABLE ... DISCARD TABLESPACE results in an assertion
being hit.

Fix
===
It should not be possible to do a flush tables for export for a
file-per-table after discarding its tablespace. This is handled by
throwing a ER_TABLESPACE_DISCARDED error even before we start
processing the request. This was fixed for non-partitioned tables
as part of fix for bug#22899690. The same fix needs to be ported
for partitioned

RB : 19889
Reviewed-by : Allen lai <zheng.lai@oracle.com>