Bug#28091735:COMMIT HANGS IF BINLOG_GROUP_COMMIT_SYNC_DELAY NOT MULTIPLE OF 1/10TH OF ITSELF
Analysis:

Setting binlog_group_commit_sync_no_delay_count option means that the
server exits the wait procedure if the number of sessions reaches the
count given before the timeout i.e is set in binlog_group_commit_sync_delay
elapses. In order to check if session count has reached the given limit,
the waiting time is broken into small intervals called as delta. The delta
is calculated as 1/10th of total wait time. After each delta waiting time
is elapsed queue size is checked to see if count has reached. If count
is not reached then wait time is modified
"to_wait = to_wait - delta".

This process is repeated until to_wait > 0 and count has not been reached.
delta is calculated like this
ulong delta=
    DBUG_EVALUATE_IF("bgc_set_infinite_delay", 100000,
                     max<ulong>(1, (to_wait * 0.1)));

When to_wait which is the value of sync_delay is a non multiple of the
delta(3) like lets say 32, as this is not a multiple of to_wait time
there will be a point in the waiting code where to_wait = 2, and then
to_wait(2) = to_wait(2) - delta(3) = -1 now as to_wait is of type ulong
so the value will be wrapped and to_wait = 18446744073709551615 [ULONG_MAX].

This will make the waiting loop continue forever.