WL#11722: Step 33 ----------------- Introduce new readKeyAttrs that simply read using normal Attrinfo. This attrinfo is fed into a new NdbPack data type DataArray that is an array of pointers and lengths where length = 0 means NULL. We also change prefix storage in T-tree to use the Attrinfo format for storing the prefix. This means an extra overhead of at least 4 bytes per column stored in the prefix and a bit more at times, but never more than 7 bytes. At the same time the efficiency in comparing is a lot higher using this method.
There is also only one prefix per node and we can win this extra
storage back most likely by increasing the size of nodes in
the T-tree.

Ensured that this method is now used for inserts and deletes in the
T-tree, for index builds during restart and for initial search of
a range scan.