WL #12074: Volcano iterator executor base [2/15, initial]
Lay down the initial framework for the iterator executor;
basic operations such as filtering and aggregation, code that
converts from the old executor's structures to iterators,
and then an actual executor.

The old executor will remain in parallel until the new can handle
every single query (which won't happen in this worklog), so that
we can fall back to it when we need to.

Change-Id: I8f29368ca11026bb646aeaff8d2072007ef6c2bf