Bug #26574003 : NODE CRASH WITH ERROR 2335 DUE TO TOO MANY DROP_TRIG_IMPL_REQ IN SQ FROM SUMA
When SUMA receives the SUB_STOP_REQ signal, it executes it and sends
back the SUB_STOP_CONF. The SUB_STOP_CONF ultimately is relayed back to
the API and the API is now open to send more SUB_STOP_REQs.

The SUMA, only after sending the SUB_STOP_CONF, asynchronously decides
to drop the subscription altogether if no subscribers are present. This
drop involves sending multiple DROP_TRIG_IMPL_REQs to DBTUP. With the
way the DbtupProxy is implemented, it can handle 21 DROP_TRIG_IMPL_REQs
in parallel. Any more subsequent requests are queued in the DbtupProxy
thread short time queue. If the execution of the DROP_TRIG_IMPL_REQ is
delayed, there is a chance where the short time queue gets overflowed
with too many DROP_TRIG_IMPL_REQs and crashes the data node with a
"Error in short time queue" error.

A similar scenario exists in SUMA node failure handling, where a node
failure can cause a lot of subscriptions to stop and can end up
overflowing the DbtupProxy with DROP_TRIG_IMPL_REQ request. This
possible overflow during node failure handling is already guarded
against by limiting the maximum DROP_TRIG_IMPL_REQs allowed to execute
parallely. But the SUB_STOP_REQ handling lacks such a guard.

This patch fixes that by delaying the execution of the SUB_STOP_REQ when
there are a number of DROP_TRIG_IMPL_REQs outstanding.
As a result,
 i.  This prevents SUB_STOP_REQ from adding 3x DROP_TRIG_IMPL_REQ to the
     DbtupProxy thread short time queue, at the cost of 1x SUB_STOP_REQ
     in the SUMA thread short time queue.
 ii. This delays sending SUB_STOP_CONF, which will stall any further
     request from the same API source, as API sources are usually
     synchronous. If multiple APIs are separately submitting
     SUB_STOP_REQs to SUMA, and are causing different subricptions to
     become STOPPED, then they will also be queued at SUMA when the
     number of DROP_TRIG_IMPL_REQs are higher than the threshold.

The threshold, maximum number of DROP_TRIG_IMPL_REQs that can be
executed parallely, is set to 6. Given that each SUB_STOP_REQ might
send 3x DROP_TRIG_IMPL_REQ to DbtupProxy and the maximum threshold set
at 6, the number of subscriptions that can be dropped in parallel within
a data node at a time is limited to 2.

Change-Id: I8879eb00af3c739b7774af5a9162427b3a40bfdd