Bug#27869124 UNIT TEST "TEST_NDBJTIE_UNLOAD_MUTILS" FAILS IN PB2
DESCRIPTION:
	Test case "MyLoadUnloadTest.java", loads a dynamic library, runs the
given test, unloads the dynamic library in a loop. At the end of each
iteration, the class that loaded the dll is forcefully garbage collected
to unload the library from test and so that it can be loaded again and
tested in next iteration. After garbage collecting the class, the memory
that dll holds were not released back to OS.

PROBLEM:
	1) As dll remains attached to the process, all global and static
variables were not reset to default. This becomes issue in case of
ndbclient lib, because ndbclient lib has some crucial variables like
std::once_flag, these variables once set can not be reset. During unload,
CharsetMap variables were un-initialized and in second iteration of test,
when the CharsetMap is re-initialized, std::once_flag is checked as it is
true, CharsetMap initialization is skipped. So, now the global pointer
points to invalid memory from first iteration and this leads to crash when
accessed.
	2) JNI_Onload() called in test main thread, has call to my_thread_init()
and JNI_OnUnLoad() which is called by garbage collector thread, has call to
my_thread_end(). But, These mysys functions where supposed to be called inside
the same thread else mutex in my_thread_global_end() timeouts and gives
error. But, this happens only in debug build.

FIX:
	Un-initialization functions where moved out of JNI_OnUnload(),and
So, it will not be called for each iteration and will be called only when
the test case gets terminated. Therefore, now there is only one init and
and only one un-init. This solves the first problem. and As per the fix now
my_thread_init() and my_thread_end() will be called in same thread.