WL#11652  Support multiple addresses for the --bind-address command option
This worklog adds support for binding a server to several IP addresses.
To specify several addresses for listening incoming connection requests,
the command line option --bind-address was extended to support comma
separated list of IP addresses to bind. For every specified IP address
a separate TCP socket is created and added to a list of sockets for
listening incoming TCP connections.