Bug#28263631 SIMULATEDBLOCK LEAK FILEOUTPUTSTREAM: COUNTINGPOOL-T DYNARR256-T INTRUSIVELIST-T
Problem
=======
In debug build the constructor of SimulatedBlock leaks an FileOutputStream
and NdbOut object used to connect its debugOut object to current output FILE
object.

  SimulatedBlock::SimulatedBlock(BlockNumber blockNumber,
                                 struct Block_context & ctx,
                                 Uint32 instanceNumber)
    : theNodeId(globalData.ownId),
    ...
      c_counterMgr(* this)
  #ifdef VM_TRACE
      ,debugOut(*new NdbOut(*new
FileOutputStream(globalSignalLoggers.getOutputStream())))
  #endif
  #ifdef VM_TRACE_TIME
      ,m_currentGsn(0)
  #endif
  {

In ASAN builds, at least three unit tests fail due to this leak is detected,
those tests are CountingPool-t, DynArr256-t, and, IntrusiveList-t.

The reason the leak is not detected in the data nodes themself is that they
never end, and if they do end it is not by returning via main function and so
leak detection is never done.

Fix
===
The extra NdbOut is not needed and can be removed.
Make the FileOutputStream member of SimulatedBlock instead of do separate
allocation of it.