Bug #27804668   CRASH DUE TO SEGMENTATION FAULT IN FIL_SPACE_IS_BEING_TRUNCATED
PROBLEM

A parent table is trying to check if a referenced table
is being truncated,but the referenced table has been
discarded. Therefore any call to find fil_space_t
structure of a discarded table causes a segmentation fault.

FIX

If the referenced table is discarded, then skip the
foreign key checks for DML.