Bug #28068293  KILLING QUERY THAT IS WAITING FOR LOCK VIA FOREIGN KEY CAUSES MYSQL CRASH
PROBLEM

Child table is waiting for a lock on the parent row in the function
lock_wait_suspend_thread() when we interrupt it with Ctrl+c. It
comes out of lock_wait_suspend_thread() with the trx->error_state
set as interrupted ,but the trx->error_state is never captured
and the error is set as DB_LOCK_WAIT, because of which we again
try to lock the row and again call lock_wait_suspend_thread(),
but since the transaction is already interrupted it comes out of
this function and again this operation continues infinitely.

FIX

Capture the trx->error_state error and pass it to upper layers
so that it is properly handled.