Bug #28239197 	NDB_IMPORT2 FAILS SPORADICALLY IN 7.6
In one of the test cases, a data node is made to fail and
ndb_import is run immediately after that.

After that a few tables are read to check for consistency.
These read statements may fail due to a temporary error caused
due to the data node shut down.

To avoid the test case from failing, this patch expects the temporary
error instead of failing.