Bug#28245522: Allocate auto releaser maps on demand
The auto releaser for dictionary objects allocated the maps for
each dictionary object type on the stack along with the allocation
of the auto releaser itself. This has a performance impact,
especially on Windows, and hence, we instead allocate the maps
dynamically on demand.

Additionally. this patch includes maps for column statistics and
resource groups in size calculations and object erasure for the
auto releaser.

Finally, in the context of executing ANALYZE TABLE, a missing
auto releaser was added.

Based on a patch written by Dan Blanchard <daniel.blanchard@oracle.com>.

Change-Id: If74760294b7ec7c36887a971aea9bdb8708bb56d