Bug#28099963 POSSIBLY UNUSED FUNCTIONS IN XDR_UTILS
Problem
-------
Clang generates warnings when compiling XCom.

Analysis
--------
We make use of the define_xdr_funcs macro even when we do not need all the
functions it generates.

Solution
--------
The community contributed a patch that marked all the functions generated by
define_xdr_funcs as unused. The patch's approach fixes the symptoms by telling
the compiler that all the functions generated by the define_xdr_funcs are always
unused. That is not true, although it does shut the compiler up.

Since the contributed patch does not apply cleanly anymore, we improve on the
submitted patch by only generating the functions that are needed.

Reviewed-by: Luis Soares <luis.soares@oracle.com>