WL#11581: Support for opening tables in a secondary storage engine
In SELECT statements where all the accessed base tables have shadow
copies in a secondary storage engine, the tables in the secondary
storage engine will be used when the statement is executed.

The secondary tables are represented by their own TABLE_SHARE and
TABLE objects in the table definition cache and table cache.

Change-Id: I5d6069f6116280a8754a57e37905c14f58110f86