Bug#28227243:WRONG ERROR TYPE RETURNED WHEN READ OF MAGIC BUFFER FAILS
Problem:

If the read of the magic buffer fails the error returned is

ERROR HY000: Error when executing command xyz: Binlog has bad magic number;
It's not a binary log file that can be used by this version of MySQL

Which is incorrect as we don't know about the magic number correctness yet.

Fix:

In the above case the read of header part which contained the magic buffer
has failed, so we should return this error. I/O error reading the header
from the binary log

Note: This bug was introduced in WL#10956