Bug #28220751 	REPLICATION BREAKS ON CREATE/DROP TEMPORARY TABLE OPERATION WITHIN FUNCTION(MBR)
Problem
=======
The F-1 of Wl#8872 states that when @@session.binlog_format=ROW/MIXED,
the server must not write CREATE TEMPORARY TABLE or DROP TEMPORARY
TABLE to the binary log. In the corner case, we just write the
'function call' top statement into the binlog as before in mixed
mode. Which didn't obey the F-1 of Wl#8872. The 'function call'
top statement is written into binary log regardless of the function
has 'CREATE/DROP TEMPORARY TABLE' or not in mixed mode, the problem
is that the 'function call' top statement is not written into binary
log if the function has only 'DROP TEMPORARY TABLE' in mixed mode.
Which causes the replication break. We have the same issue in
statement mode.

Analyze
=======
The 'function call' top statement can not distinguish if its
sub statements (function) have 'CREATE/DROP TEMPORARY TABLE'
or not before executing its sub statements, so setting the
binlog format to row in mixed mode before executing its sub
statements for all cases causes many issues. I can see many
tests failures with the solution. It is too late to set the
binlog format to row in mixed mode when executing the
'CREATE TEMPORARY TABLE' in sub statement, because the binlog
format is not consistent before and after 'CREATE TEMPORARY TABLE'.

Fix
===
To fix the problem, write a 'function call' top statement into
binary log even if the function has only 'DROP TEMPORARY TABLE'
as others in mixed and statement mode.