Bug #17732772 NDB_MGMD : CLUSTER LOG SYNCHRONOUSLY WRITTEN BY SIGNAL HANDLING THREAD
Code refactoring:
Pass timestamp to LogHandler::append(), don't store it as a member.
Pass it down to getDefaultHandler() and make getDefaultHeader() use this for
header generation.

This change is made so that all the LogHandler::append() parameters can be
stored in a buffer and retrieved later and reconstitued to generate the log message.