Bug #17732772 NDB_MGMD : CLUSTER LOG SYNCHRONOUSLY WRITTEN BY SIGNAL HANDLING THREAD
Introduce a new LogHandler type- BufferedLogHandler.
This type of log handler stores all the required data to create a log message in a
log buffer stored in memory (object of type LogBuffer).

It also creates a logging thread which reads log messages from the log buffer and logs
them to a destination log handler.