Bug #17732772 NDB_MGMD : CLUSTER LOG SYNCHRONOUSLY WRITTEN BY SIGNAL HANDLING THREAD
Make cluster logging asynchronous when LogDestination is set to FILE.
All the log messages intended to be written into the cluster log file are
redirected to a BufferedLogHandler object which stores the log data in a
LogBuffer object.
Internally, the BufferedLogHandler object creates a log thread which reads
from the log buffer and uses a FileLogHandler object to write the log messages
to the cluster log file using the destination file log handler and the
retrieved log information.

Note: Logging is made asynchronous only in the worker process and not
in the parent.