Bug#26876608 MYSQL CRASH:CACHE_POLICY=CACHING|INNODB_ONLY & SET GLOBAL INNODB_API_TRX_LEVEL=1
PROBLEM:
Assert in innodb with innodb_api_trx_level=1 for
the case when the transaction is not committed in SE
and memcached api tries to read that.

ANALYSIS:
in row_search_mvcc function, there is check if
mysql_handler->end_range is not null.
for the memcached transaction, mysql_handler is not
set.

FIX:
put addition check if the transaction is started from
memcached or not.

REVIEWED BY: Jimmy Yang<Jimmy.Yang@oracle.com>
RB: 19922