WL#11879 Replicate original server version
This worklog implements two new session variables:
original_server_version and immediate_server_version.
Original_server_version is used to tag each transaction with the version
of the server where it was originally executed, more specifically, with the
value of @@GLOBAL.version. This variable is replicated, unchanged, to the
servers in the replication topology.
Immediate_server_version is used to tag each transaction with the
version of the immediate master in the replication topology.