WL#11722: Step 44 ----------------- When sending the CONTINUEB to restart a scan after a queueing event, it was important to handle the release of this scan record properly. We need to keep the record until this signal arrives. So when in the state, waiting for restart of scan it is important to not release the record. This was previously handled in finishScanrec and immediately thereafter we called releaseScanrec to set the variables of the scan record properly.
Actually this is not correct, this means we first release the
record and then update it. This could create problems in future
poolification efforts. So now it is not released until all
updates of the record have completed. Also it is not released
until we are sure that there are no outstanding signals to this
scan record.

We put the finish scan code into a new method called handle_finish_scan.
Copy scans don't really need this code since they can never be in a
queue, but it was still seen as easier to make this code in one
method since there are so much code reuse.